Separated game mechanics into Snake_Game.cpp
and updated makefile and stuff
