+++ snake/makefile
@@ -1,8 +1,8 @@
 play: snake
 	./snake
 
-snake: snake_game.cpp
-	g++ snake_game.cpp Game.cpp -lncurses -O1 -o snake
+snake: snake.cpp
+	g++ snake.cpp Screen.cpp Snake_Game.cpp -lncurses -O1 -o snake
 
 clean:
 	\rm snake
